requirements.txt
aioredis==1.3.1
alembic==1.5.8
altgraph==0.17
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
auto-py-to-exe==2.7.11
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
blinker==1.4
bottle==0.12.19
bottle-websocket==0.2.9
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
channels==2.4.0
channels-redis==2.4.2
chardet==4.0.0
click==7.1.2
colorama==0.4.4
comtypes==1.1.8
constantly==15.1.0
cryptography==3.4.7
cycler==0.10.0
daphne==2.5.0
decorator==5.0.5
Django==3.1.5
django-ckeditor==6.0.0
django-js-asset==1.2.2
dnspython==2.1.0
docutils==0.16
dominate==2.6.0
EasyProcess==0.3
Eel==0.12.4
email-validator==1.1.2
Flask==1.1.2
Flask-Admin==1.5.8
Flask-BabelEx==0.9.4
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.4.1
Flask-Ext==0.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
graphviz==0.16
greenlet==1.0.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
infinity==1.5
intervals==0.9.1
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-examples==2.0.0
Kivy-Garden==0.1.4
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lxml==4.6.2
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
MouseInfo==0.1.3
msgpack==0.6.2
multi-clipboard==1.3.0
mysql-connector==2.2.9
mysqlclient==2.0.3
neat==0.4.1
neat-python==0.92
numpy==1.19.5
opencv-python==4.5.1.48
pandas==1.2.1
passlib==1.7.4
pefile==2019.4.18
phonenumbers==8.12.20
Pillow==8.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycparser==2.20
pygame==2.0.1
PyGetWindow==0.0.9
Pygments==2.7.4
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pyjokes==0.6.0
PyMsgBox==1.0.9
pynput==1.7.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.1
pypiwin32==223
PyQt5==5.15.2
PyQt5-sip==12.8.1
PyQt5Designer==5.14.1
PyQtWebEngine==5.15.2
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
python-dotenv==0.17.0
python-editor==1.0.4
pyttsx3==2.90
PyTweening==1.0.3
pytz==2020.5
PyVirtualDisplay==2.0
pywhatkit==3.9
pywin32==300
pywin32-ctypes==0.2.0
redis==3.5.3
requests==2.25.1
scipy==1.6.1
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.1
speaklater==1.3
SpeechRecognition==3.8.1
SQLAlchemy==1.4.7
SQLAlchemy-Utils==0.36.8
sqlparse==0.4.1
tkcalendar==1.6.1
toml==0.10.2
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
urllib3==1.26.2
validators==0.18.2
visitor==0.1.3
watchdog==1.0.2
WebOb==1.8.6
Werkzeug==1.0.1
whichcraft==0.6.1
wikipedia==1.4.0
wrapt==1.12.1
WTForms==2.3.3
WTForms-Alchemy==0.17.0
WTForms-Components==0.10.5
zope.event==4.5.0
zope.interface==5.2.0
